Sticky footer and bootstrap tab adjustments
I added the “Site-content” class back to make the footer sticky and
added it to the home page to fix the Safari bug which made the home
page background image disappear.

I also removed the margin from the bootstrap tabs on mobile which
caused the first tab to be larger.